float -> bigdecimal & total price logic

+++ src/main/java/com/example/whynotpc/models/dto/OrderDTO.java
@@ -2,12 +2,13 @@ package com.example.whynotpc.models.dto;
 
 import com.example.whynotpc.models.order.OrderItem;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderDTO(
         Integer id,
         String status,
-        Float total,
+        BigDecimal total,
         List<OrderItem> items,
         Integer userId
 ) {
